requirements.txt
colorama
numpy
pillow
suya
tableprint
tensorboard
torch
torchvision
torchsummary
tqdm